% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{somme}
\alias{somme}
\alias{multiplication}
\title{Calculs basiques}
\usage{
somme(x, y)

multiplication(x, y)
}
\arguments{
\item{x}{Numerique.}

\item{y}{Numerique.}
}
\value{
La somme ou le produit de \code{x} et \code{y}
}
\description{
Calculs basiques
}
\examples{
somme(4,7)
}
